Run OONI v2.2.0 from debian 8 (#6)
This loosely matches what happens with lepidopter. We started
distributing lepidopter in the summer of 2016. We were certainly
using Debian 8 at the time. Also, we were probably using a version
of OONI that was lesser than v2.2.0, which was released on April
2017. But there is also auto-updating, so it's not fully clear to
me what is the right approach here.

Another aspect to consider is that I have pinned all the dependencies
to the dependencies that are used by OONI v2.3.0, because not pinning in
a precise way the depedencies is going to cause failures.

That is, the pursuit of the correct combination of dependencies that
faithfully represent lepidtoper is not so easy.

(It's true that there is a list of pinned dependencies at
github.com/TheTorProject/lepidopter but these dependencies
are not sufficient to avoid build errors.)

This is part of https://github.com/ooni/backend/issues/446.