config/ooniprobe220pip: reduce gap with lepidopter packages
In this diff I'm trying to reduce the diff between the packages
installed by lepidopter and the ones we install.

I've reached the point where the diff seems reasonable and I think
a couple of packages should actually not be the ones used by lepidopter
since that would be a downgrade.

Part of https://github.com/ooni/backend/issues/446#issuecomment-697172389

The following is the diff between lepidopter and us:

```
--- lepidopter-v1.0.0-armel-pip-packages	2020-09-23 14:10:15.000000000 +0200
+++ config/ooniprobe220pip/requirements.txt	2020-09-23 17:00:33.000000000 +0200
@@ -4,6 +4,7 @@
 PyYAML==3.12
 Twisted==17.1.0
 Werkzeug==0.12.1
+appdirs==1.4.3
 asn1crypto==0.22.0
 attrs==16.3.0
 certifi==2017.4.17
@@ -13,14 +14,13 @@
 enum34==1.1.6
 fte==0.1.3
 fteproxy==0.2.19
-idna==2.2
+idna==2.5
 incremental==16.10.1
 ipaddr==2.1.11
-ipaddress==1.0.17
+ipaddress==1.0.18
 keyrings.alt==1.3
 klein==17.2.0
 obfsproxy==0.2.13
-ooniprobe==2.2.0
 packaging==16.8
 pyOpenSSL==17.0.0
 pyasn1-modules==0.0.8
@@ -30,11 +30,12 @@
 pydumbnet==1.12.1
 pygobject==3.22.0
 pyparsing==2.2.0
-pypcap==1.1.6
+pypcap==1.2.1
 pyptlib==0.0.6
 pyxdg==0.25
 scapy==2.3.3
 service-identity==16.0.0
+setuptools==11.3
 six==1.10.0
 txsocksx==1.15.0.2
 txtorcon==0.18.0
```