config/ooniprobe220pip: reduce gap with lepidopter packages (#13)
In this diff I'm trying to reduce the diff between the packages
installed by lepidopter and the ones we install.

I've reached the point where the diff seems reasonable and I think
a couple of packages should actually not be the ones used by lepidopter
since that would be a downgrade.

Part of https://github.com/ooni/backend/issues/446#issuecomment-697172389

The following is the diff between lepidopter and us:

```
--- lepidopter-v1.0.0-armel-pip-packages	2020-09-23 18:07:49.000000000 +0200
+++ config/ooniprobe220pip/requirements.txt	2020-09-23 18:08:23.000000000 +0200
@@ -3,39 +3,31 @@
 Parsley==1.3
 PyYAML==3.12
 Twisted==17.1.0
-Werkzeug==0.12.1
-asn1crypto==0.22.0
+Werkzeug==0.12
+appdirs==1.4.3
+asn1crypto==0.21.1
 attrs==16.3.0
-certifi==2017.4.17
-cffi==1.10.0
+certifi==2017.1.23
+cffi==1.9.1
 constantly==15.1.0
 cryptography==1.8.1
 enum34==1.1.6
-fte==0.1.3
-fteproxy==0.2.19
-idna==2.2
+idna==2.5
 incremental==16.10.1
 ipaddr==2.1.11
-ipaddress==1.0.17
-keyrings.alt==1.3
+ipaddress==1.0.18
 klein==17.2.0
-obfsproxy==0.2.13
-ooniprobe==2.2.0
 packaging==16.8
-pyOpenSSL==17.0.0
+pyOpenSSL==16.2.0
 pyasn1-modules==0.0.8
 pyasn1==0.2.3
 pycparser==2.17
-pycrypto==2.6.1
 pydumbnet==1.12.1
-pygobject==3.22.0
 pyparsing==2.2.0
-pypcap==1.1.6
-pyptlib==0.0.6
-pyxdg==0.25
+pypcap==1.2.1
 scapy==2.3.3
 service-identity==16.0.0
-six==1.10.0
+setuptools==11.3
 txsocksx==1.15.0.2
 txtorcon==0.18.0
-zope.interface==4.4.0
+zope.interface==4.3.3
```

So, we're not exactly using the same packages but hopefully we're close enough that we can say "hey, this package is having some form of lepidopter like experience". I may change this in future as I evaluate other ideas. Meanwhile, it's good to merge this change because it definitely starts to represent some form of progress.