Start probing the testing cloudfronted endpoint
Because it's not immediately obvious to me what is the correct way to
force a cloudfronted bouncer, I am using a simple strategy for now where
we treat the cloudfronted URL as a normal HTTPS URL.

What are we testing here? Well, I would say that we're checking whether
different incarnations of the probe can speak to the cloudfronted backend.

We're not testing MK because it does not support cloudfronting.

Part of https://github.com/ooni/backend/issues/451.