fix: tweak legacy probe versions and targets
Limit the runs to versions of the legacy probes that have been
running non-ndt-measurements in the last seven days.

See https://github.com/ooni/backend/issues/461#issuecomment-726587481.

Make sure we target both production and testing.

As far as testing is concerned, just point to the HTTPS version to
ensure that there are no API breakages for now.